gsup: add OSMO_GSUP_RAT_TYPES_IE
Corresponds to the Supported RAT Types Indicator from 3GPP TS 29.002. See 8.1.2
MAP_UPDATE_LOCATION service.

This is useful to communicate the RAN / RAT type of the current subscriber
during Location Updating Request.

The use case we have in mind for VLR <-> HLR would require only a single RAT
type, but 29.002 specifies a list of supported RAT types. Hence implement this
as a list.

So far, have room for eight RAT types in the gsup_msg. That is an arbitrary
random choice without any rationale.

Change-Id: I93850710ab55a605bf61b95063a69682a2899bb1