doxygen: allow manual 'make apidoc' even if --disable-doxygen
Even if ./configure --disable-doxygen was passed and doxygen builds are by
default skipped, provide a manual 'make apidoc' target that nevertheless
generates the API doc on the premise that a 'doxygen' program is available.

Especially since we do a two-pass doxygen build whenever any source file
changes, my guess is --disable-doxygen could be a common choice. It is then
cumbersome to have to ./configure just to get one doxygen build started.

Change-Id: If8d8dfb8365c8f28612b8ce2b8ddf88f74df9a90