make use of OTC_GLOBAL when allocating library-internal contexts
As libosmcore is now managing the global talloc contexts, there's
no point in having APIs where the user tells the library about
which talloc contexts to use for a given sub-system.

However, completely ignoring the talloc contexts passed in by
existing applications would unfortunately break LeakSanitizer.  It
appears to track if any dynamically allocated pointeres are not stored
anywhere, and that would be exactly the situation created here.

So the new behavior is: Use the context passed in by an application,
and fall back to the library-internal OTC_GLOBAL if none is passed in.

Change-Id: I48f475efd3ee0d5120b8fc30861e852d1a6920b1